update and finish manger and add main that run app
